Finish 3.0-alpha.1
