+++ build.gradle
@@ -1,5 +1,5 @@
 plugins {
-    id 'fabric-loom' version '0.10.9'
+    id 'fabric-loom' version '0.10.50'
     id 'maven-publish'
 }
