+++ gradle.properties
@@ -3,12 +3,12 @@ org.gradle.jvmargs=-Xmx1G
 # Fabric Properties
 # check these on https://modmuss50.me/fabric.html
 minecraft_version=1.17.1
-yarn_mappings=1.17.1+build.61
-loader_version=0.11.7
+yarn_mappings=1.17.1+build.63
+loader_version=0.12.5
 # Mod Properties
-mod_version=2.2-pre.1
+mod_version=3.0-alpha.1
 maven_group=com.redgrapefruit.itemnbt
 archives_base_name=itemnbt
 # Dependencies
 # check this on https://modmuss50.me/fabric.html
-fabric_version=0.40.1+1.17
+fabric_version=0.42.0+1.17
